binance_momentum_monitor/requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
requests==2.31.0
pyyaml==6.0.1
websockets==12.0
numpy==1.26.2
